--- learn2rag/pipeline/__main__.py
@@ -1,7 +1,0 @@
-import logging
-
-from . import main
-
-logging.basicConfig(level=logging.WARNING)
-
-main()
